Refactor File and CSVFile to extend from FileBase superclass (TBD path setting from designer in CSVFile
